# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Estude Loterias API - Estatísticas e ferramentas para loterias da Caixa
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /maismilionaria/resultados:
    get:
      operationId: getMaismilionariaResultados
      tags: [maismilionaria]
      summary: Resultados paginados da +Milionária
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 50
            default: 20
        - name: ano
          in: query
          schema:
            type: integer
            nullable: true
        - name: ordem
          in: query
          schema:
            type: string
            enum: [asc, desc]
            default: desc
      responses:
        "200":
          description: Resultados paginados
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResultadosPaginadosMaismilionaria"
  /maismilionaria/resultados/{concurso}:
    get:
      operationId: getMaismilionariaResultadoConcurso
      tags: [maismilionaria]
      summary: Resultado de um concurso específico da +Milionária
      parameters:
        - name: concurso
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Resultado do concurso
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResultadoMaismilionaria"
        "404":
          description: Concurso não encontrado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
...
  /maismilionaria/simulador:
    post:
      operationId: simularMaismilionaria
      tags: [maismilionaria]
      summary: Simula uma aposta no histórico da +Milionária
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SimuladorInputMaismilionaria"
      responses:
        "200":
          description: Resultado da simulação
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SimulacaoResultado"
  /maismilionaria/gerador:
    post:
      operationId: gerarJogoMaismilionaria
      tags: [maismilionaria]
      summary: Gera um jogo aleatório de +Milionária
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GeradorInputMaismilionaria"
      responses:
        "200":
          description: Jogo(s) gerado(s)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JogoGerado"
components:
  schemas:
...
    ApiError:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
    FaixaPremio:
      type: object
      properties:
        faixa:
          type: integer
        descricao:
          type: string
        ganhadores:
          type: integer
        valorPremio:
          type: number
      required:
        - faixa
        - descricao
        - ganhadores
        - valorPremio
...
    SimulacaoResultado:
      type: object
      required: [resumo, concursos, totalPremio, totalConcursos]
      properties:
        resumo:
          type: array
          items:
            type: object
            required: [acertos, contagem]
            properties:
              acertos:
                type: integer
              contagem:
                type: integer
        resumoTrevos:
          type: array
          items:
            type: object
            required: [acertos, contagem]
            properties:
              acertos:
                type: integer
              contagem:
                type: integer
        faixas:
          type: array
          items:
            type: object
            required: [label, contagem, probabilidade]
            properties:
              label:
                type: string
              contagem:
                type: integer
              probabilidade:
                type: string
        concursos:
          type: array
          items:
            type: object
            required: [concurso, data, dezenas, acertos, premioGanho]
            properties:
              concurso:
                type: integer
              data:
                type: string
              dezenas:
                type: array
                items:
                  type: string
              trevos:
                type: array
                items:
                  type: string
              acertos:
                type: integer
              acertosTrevos:
                type: integer
              premioGanho:
                type: number
        totalPremio:
          type: number
        totalConcursos:
          type: integer
...
    JogoGerado:
      type: object
      properties:
        jogos:
          type: array
          items:
            type: array
            items:
              type: integer
        custo:
          type: number
      required:
        - jogos
        - custo
...
    ResultadoMaismilionaria:
      type: object
      properties:
        concurso:
          type: integer
        data:
          type: string
        dezenas:
          type: array
          items:
            type: string
        dezenasOrdem:
          type: array
          nullable: true
          items:
            type: string
        trevos:
          type: array
          nullable: true
          items:
            type: string
        trevosOrdem:
          type: array
          nullable: true
          items:
            type: string
        premios:
          type: array
          items:
            $ref: "#/components/schemas/FaixaPremio"
        acumulado:
          type: boolean
        valorAcumulado:
          type: ["number", "null"]
        dataProximoConcurso:
          type: ["string", "null"]
        valorEstimadoProximoConcurso:
          type: ["number", "null"]
        arrecadacaoTotal:
          type: ["number", "null"]
      required:
        - concurso
        - data
        - dezenas
        - premios
        - acumulado
    ResultadosPaginadosMaismilionaria:
      type: object
      properties:
        total:
          type: integer
        pagina:
          type: integer
        limite:
          type: integer
        totalPaginas:
          type: integer
        resultados:
          type: array
          items:
            $ref: "#/components/schemas/ResultadoMaismilionaria"
      required:
        - total
        - pagina
        - limite
        - totalPaginas
        - resultados
...
    SimuladorInputMaismilionaria:
      type: object
      required:
        - dezenas
        - trevos
      properties:
        dezenas:
          type: array
          items:
            type: string
          minItems: 6
          maxItems: 12
        trevos:
          type: array
          items:
            type: string
          minItems: 2
          maxItems: 6
        filtro:
          type: string
          enum:
            - todos
            - premiados
            - seis
            - cinco
            - quatro
            - tres
          default: premiados
    GeradorInputMaismilionaria:
      type: object
      properties:
        quantidadeJogos:
          type: integer
          minimum: 1
          maximum: 20
          default: 1
        quantidadeDezenas:
          type: integer
          minimum: 6
          maximum: 12
          default: 6
        quantidadeTrevos:
          type: integer
          minimum: 2
          maximum: 6
          default: 2
      required:
        - quantidadeJogos