# lib/api-zod/src/generated/types/resultadoMaismilionaria.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Estude Loterias API - Estatísticas e ferramentas para loterias da Caixa
 * OpenAPI spec version: 0.1.0
 */
import type { FaixaPremio } from './faixaPremio';

export interface ResultadoMaismilionaria {
  concurso: number;
  data: string;
  dezenas: string[];
  dezenasOrdem?: string[] | null;
  trevos?: string[] | null;
  trevosOrdem?: string[] | null;
  premios: FaixaPremio[];
  acumulado: boolean;
  /** @nullable */
  valorAcumulado?: number | null;
  /** @nullable */
  dataProximoConcurso?: string | null;
  /** @nullable */
  valorEstimadoProximoConcurso?: number | null;
  /** @nullable */
  arrecadacaoTotal?: number | null;
}